backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
pillow
opencv-python
pydantic
pandas
numpy
